(PC-1543) Configure only offerer view
