+++ admin/custom_views.py
@@ -1,0 +1,10 @@
+from admin.base_configuration import BaseAdminView
+
+
+class OffererAdminView(BaseAdminView):
+    can_edit = True
+    column_list = ['id', 'name', 'siren', 'city', 'postalCode', 'address']
+    column_labels = dict(name='Nom', siren='SIREN', city='Ville', postalCode='Code postal', address='Adresse')
+    column_searchable_list = ['name', 'siren']
+    column_filters = ['postalCode', 'city']
+    form_columns = ['name', 'siren', 'city', 'postalCode', 'address']
