+++ templates/admin/index.html
@@ -2,8 +2,8 @@
 
 {% block body %}
     {% if current_user.is_authenticated and current_user.isAdmin %}
-        <div class="alert alert-success" role="alert">
-            Bienvenue dans l'admin de pass Culture !
+        <div class="alert alert-info" role="alert">
+            Bonjour <strong>{{ current_user.publicName }}</strong>, bienvenue dans l'interface d'administration du pass Culture !
         </div>
     {% else %}
         <div class="alert alert-danger" role="alert">
